requirements.txt
affine @ file:///Users/ktietz/demo/mc3/conda-bld/affine_1629890479180/work
aiohttp==3.9.5
aiosignal==1.3.1
asciitree==0.3.3
asgiref @ file:///Users/ec2-user/ci_py311/asgiref_1678390266874/work
attrs @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_8d7cifg9jt/croot/attrs_1695717834289/work
boto3==1.34.142
botocore==1.34.142
Bottleneck @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_cfe9itj_y_/croot/bottleneck_1707864223016/work
cachier==3.0.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_73k3qumdla/croot/certifi_1720453498134/work/certifi
cftime==1.6.4
charset-normalizer==3.3.2
click @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_f27dk3zwrb/croot/click_1698129816363/work
click-plugins @ file:///Users/ktietz/demo/mc3/conda-bld/click-plugins_1630665883852/work
cligj==0.7.2
cloudpickle==3.0.0
contourpy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_deuzpvbq4p/croot/contourpy_1700583585680/work
copernicusmarine==1.3.1
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dask==2024.7.0
Django @ file:///Users/ec2-user/ci_py311/django_1678411010555/work
fasteners==0.19
fonttools @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_801fru8lup/croot/fonttools_1713551493489/work
frozenlist==1.4.1
fsspec==2024.6.1
idna==3.7
importlib_metadata==8.0.0
jmespath==1.0.1
kiwisolver @ file:///Users/ec2-user/ci_py311/kiwisolver_1678322457192/work
locket==1.0.0
lxml==5.2.2
matplotlib @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_56lkmgsho5/croot/matplotlib-suite_1713336381321/work
mkl-fft @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_af8wl3mjfu/croot/mkl_fft_1695058185839/work
mkl-random @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_0fq2s894lk/croot/mkl_random_1695059822529/work
mkl-service==2.4.0
multidict==6.0.5
nest-asyncio==1.6.0
netCDF4==1.7.1.post1
numcodecs==0.12.1
numexpr @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_c2pokmr868/croot/numexpr_1696515300257/work
numpy @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_b1_7dulbi4/croot/numpy_and_numpy_base_1695830449184/work/dist/numpy-1.26.0-cp311-cp311-macosx_10_9_x86_64.whl#sha256=df7810f057779b4c06be5bdbd7924c5dc5b91a218ad44586e9893864b3ebaae4
packaging @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_af6w2d9k11/croot/packaging_1720101860461/work
pandas @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_f3c5inakly/croot/pandas_1718308972943/work/dist/pandas-2.2.2-cp311-cp311-macosx_10_15_x86_64.whl#sha256=b0a3da9b3c56b4738820f67ec891e7b0e38e29cf7df56efa417f584270ef0e79
partd==1.4.2
pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1719903579516/work
portalocker==2.10.0
pyparsing @ file:///Users/ec2-user/ci_py311/pyparsing_1678322828710/work
pystac==1.10.1
python-dateutil @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_efk5_uakg8/croot/python-dateutil_1716495742183/work
pytz @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_244ln7biq9/croot/pytz_1713974320727/work
PyYAML==6.0.1
rasterio @ file:///Users/runner/miniforge3/conda-bld/rasterio_1717673082977/work
requests==2.32.3
s3transfer==0.10.2
semver==3.0.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
snuggs @ file:///Users/ktietz/demo/mc3/conda-bld/snuggs_1629803886966/work
sqlparse @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_cb77ukgrpu/croot/sqlparse_1690901947630/work
toolz==0.12.1
tornado @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_7e0ruwlfee/croot/tornado_1718740118099/work
tqdm==4.66.4
tzdata @ file:///croot/python-tzdata_1690578112552/work
unicodedata2 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_1anajuo38r/croot/unicodedata2_1713212953928/work
urllib3==2.2.2
watchdog==4.0.1
xarray @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_ba3w2rvoxd/croot/xarray_1689041477488/work
yarl==1.9.4
zarr==2.18.2
zipp==3.19.2
